PGM sender and receiver changed to new command processing interface
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@915 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libzmq/pgm_socket.cpp
@@ -306,7 +306,7 @@ size_t zmq::pgm_socket_t::write_one_pkt (unsigned char *data_, size_t data_len_)
     
     // We have to write all data as one packet.
     if (nbytes > 0) {
-        assert (nbytes == data_len_);
+        assert (nbytes == (ssize_t)data_len_);
     }
 
     return nbytes;
